fix  deployment error
